Created generic traits module
Created a generic traits module to hold all traits which aren't necessarily
for integers or floating point numbers.
