+++ src/lib.rs
@@ -4,8 +4,9 @@ pub mod traits;
 #[rustfmt::skip]
 pub mod prelude {
     pub use crate::{
+        traits::*,
         traits::float::*,
+        traits::generic::*,
         traits::integer::*,
-        traits::*
     };
 }
